upload Github data (partial implementation)

+++ models/repository.go
@@ -26,6 +26,6 @@ type TrackerRepository interface {
 
 // TrackerQueryRepository encapsulate storage & retrieval of tracker queries
 type TrackerQueryRepository interface {
-	Create(ctx context.Context, query string, schedule string, tracker int) (*app.TrackerQuery, error)
+	Create(ctx context.Context, query string, schedule string, tracker uint64) (*app.TrackerQuery, error)
 	Save(ctx context.Context, tq app.TrackerQuery) (*app.TrackerQuery, error)
 }
